Move TimeSeriesMetadataIndex to timeseries package
