+++ metrix-mapping/src/main/java/com/powsybl/metrix/mapping/timeseries/TimeSeriesMetadataIndex.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.metrix.mapping.exception;
+package com.powsybl.metrix.mapping.timeseries;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
